requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
beartype==0.18.5
billiard==4.2.0
cachetools==5.3.3
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
fast_bitrix24==1.7.3
frozenlist==1.4.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
gspread==6.1.2
httplib2==0.22.0
icontract==2.6.6
idna==3.7
kombu==5.3.7
more-itertools==10.2.0
multidict==6.0.5
oauth2client==4.1.3
oauthlib==3.2.2
prompt_toolkit==3.0.45
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.4
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
tqdm==4.66.4
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
